requirements.txt
accessify==0.3.1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
pycryptodome==3.15.0
requests==2.28.1
requests-futures==1.0.0
urllib3==1.26.12